+++ workers/unity/Assets/Playground/Scripts/Metrics/MetricSendSystem.cs
@@ -6,6 +6,7 @@ using UnityEngine;
 
 namespace Playground
 {
+    [DisableAutoCreation]
     public class MetricSendSystem : ComponentSystem
     {
         private WorkerSystem worker;
@@ -38,7 +39,7 @@ namespace Playground
                 ? DefaultTargetFrameRate
                 : Application.targetFrameRate;
 
-            lastFrameCount = Time.frameCount;
+            lastFrameCount = UnityEngine.Time.frameCount;
             calculatedFps = targetFps;
 
             timeOfLastUpdate = DateTime.Now;
@@ -71,8 +72,8 @@ namespace Playground
 
         private void CalculateFps()
         {
-            var frameCount = Time.frameCount - lastFrameCount;
-            lastFrameCount = Time.frameCount;
+            var frameCount = UnityEngine.Time.frameCount - lastFrameCount;
+            lastFrameCount = UnityEngine.Time.frameCount;
             var rawFps = frameCount / (DateTime.Now - timeOfLastUpdate).TotalSeconds;
             calculatedFps = (rawFps * (1 - smoothing)) + (calculatedFps * smoothing);
         }