+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Breaking Changes
 
 - The minimum version of Unity that the GDK supports is now 2019.3. You need to upgrade your Unity project. [#1336](https://github.com/spatialos/gdk-for-unity/pull/1336)
-- Upgraded the `com.unity.entities` packages to `0.3.0-preview.4`. [#1337](https://github.com/spatialos/gdk-for-unity/pull/1337)
+- Upgraded the `com.unity.entities` packages to `0.7.0-preview.19`. [#1340](https://github.com/spatialos/gdk-for-unity/pull/1340)
     - Authority in ECS is no longer a `SharedComponentData`, but instead a tag component.
     - `ComponentAuthority` has been renamed to `HasAuthority`.
     - No need to filter your ECS query anymore, instead match for `HasAuthority`.